Updating to typescript
This is just a dirty job to get typescript into the project, so that future contributions can begin type-annotations.
