+++ src/moduleDeclaration.d.ts
@@ -1,0 +1,1 @@
+declare module "reportWebVitals";
